Update README.md
change all the documentation examples to encourage users pass the real values instead of var=value. because there are some tiny bugs where param vars are differ from function to function and this is almost always cause a confusion. passing the values with quotes (like 'the-actual-table-name') is always easy and prevents mistakes