fix: table.rename() prefix new table name with current schema
The API for renaming a table requires fully qualified new name (schema
prefix). If it is not given, the table is renamed but also moved to the
empty/global schema.
Here we change add the schema prefix to the API call so the table will
be renamed but remain it it's current schema.