query data : extract state out of parser and fields
We create a parser for every subsplit in every response, when the table
has a lot of columns, those creations become the dominate factor in CPU
time. This change extracts the state out of the parser so it can be
created once and operate on multiple states, creating a state is
independent of the table size.