Elysium DDL support for the pysdk
`create_table()` has a new parameter, `sorting_key` to specify the sorted
columns via column index (0-based).  This parameter is silently ignored if
Elysium is not supported.

`Table` has a new method, `sorted_columns()` which returns the metadata of the
sorted columns in the order they apepar in the sorting key.  This function will
return an empty list if Elysium is not supprted.

`Table` has a new method, `add_sorting_key()` To turn an unsorted table into a
sorted one.  This method is effectilvely a no op, if Elysium is not supported.
Adding sorting key to an already sorted table results in a
`vastdb.errors.BadRequest` exception