VSDK-34: Block table creation and modification of unsupported schema types by Ibis
It seems that Ibis does not support Arrow.FixedSizeBinaryType at the moment.
Attempting to create such a table previously succeeded, but an attempt to "parse" it into a Table object
fails when converting the arrow schema into ibis.

Until Ibis adds support it is important to block usages of unsupported types with the PySDK.