ORION-278801: Fix test_vectors usage with inclusive random.randint
Since random.randint is inclusive, when it returned the num_rows the test failed since the id is up to num_rows-1.
Better to use random.choice it matches better what we would like to do.

Regarding test_select_stop the query id probably supported up to uint64 but I guess the intention was for it to be up to UINT32_MAX which is 2**32 - 1.